refactor: move profile related entities to a folder
